Preprocessed MERFISH dataset of the mouse preoptic area for a bregma -0.29 slice from a female naive animal (Animal ID = 1, Animal Sex = "Female", Behavior = "Naive", Bregma = "-0.29").

Preprocessed MERFISH dataset of the mouse preoptic area for a bregma -0.29 slice from a female naive animal (Animal ID = 1, Animal Sex = "Female", Behavior = "Naive", Bregma = "-0.29").

Usage

data("merfish_mousePOA")

Format

SpatialExperiment object where assay slot contains genes-by-cells matrix with preprocessed gene expression (total RNA counts per cell divided by cell volume and scaled by 1000) as dgCMatrix, spatialCoords slot contains x,y coordinates of cells, and colData slot contains bregma, cell type, and neuron type meta data.

Source

https://www.science.org/doi/10.1126/science.aau5324

Value

SpatialExperiment object for the preprocessed MERFISH dataset of the mouse preoptic area for a bregma -0.29 slice from a female naive animal (Animal ID = 1, Animal Sex = "Female", Behavior = "Naive", Bregma = "-0.29").